requirements.txt
crewai>=0.28.0
crewai-tools>=0.12.0
python-dotenv>=1.0.0
requests>=2.31.0
pydantic>=2.0.0
PyGithub>=2.1.1
pyyaml>=6.0